Remove some pedantic warnings

+++ client/client_test.cc
@@ -2131,7 +2131,7 @@ TEST_F(ClientTest, Mikael) {
   }
 
   ASSERT_EQ(sent_msgs.size(), received_msgs.size());
-  for (int i = 0; i < sent_msgs.size(); i++) {
+  for (size_t i = 0; i < sent_msgs.size(); i++) {
     EXPECT_EQ(sent_msgs[i], received_msgs[i]) << "i = " << i;
   }
 }
